moved all functions from ssh-ksc to library
