+++ app/ssh-ksc/ssh-ksc.go
@@ -8,15 +8,11 @@ import (
 	"fmt"
 	"github.com/codegangsta/cli"
 	"github.com/dedis/cothority/lib/dbg"
-	"github.com/dedis/cothority/lib/network"
 	"github.com/dedis/cothority/lib/ssh-ks"
 	"os"
 )
 
-func init() {
-	network.RegisterMessageType(ssh_ks.Config{})
-}
-
+// Our clientApp configuration
 var clientApp *ssh_ks.ClientApp
 
 func main() {
@@ -104,15 +100,6 @@ func serverAdd(c *cli.Context) {
 	ServerAdd(clientApp, srvAddr)
 }
 
-func ServerAdd(ca *ssh_ks.ClientApp, srvAddr string) {
-	srv, err := ssh_ks.NetworkGetServer(srvAddr)
-	dbg.ErrFatal(err)
-	err = ca.NetworkAddServer(srv)
-	dbg.ErrFatal(err)
-	conf, err := ca.NetworkSign(srv)
-	dbg.ErrFatal(err)
-	ca.Config = conf
-}
 func serverDel(c *cli.Context) {
 	srvAddr := c.Args().First()
 	dbg.Print("Deleting server", srvAddr)
@@ -123,64 +110,25 @@ func serverDel(c *cli.Context) {
 	}
 }
 
-func ServerDel(ca *ssh_ks.ClientApp, srvAddr string) error {
-	srv, err := ssh_ks.NetworkGetServer(srvAddr)
-	dbg.ErrFatal(err)
-	err = ca.NetworkDelServer(srv)
-	dbg.ErrFatal(err)
-	if len(ca.Config.Servers) == 1 {
-		dbg.Lvl2("Deleted last server")
-		ca.Config = ssh_ks.NewConfig(0)
-	} else {
-		for _, s := range ca.Config.Servers {
-			dbg.Lvl3(s.Entity.Addresses[0], srv.Entity.Addresses[0])
-			if s.Entity.Addresses[0] != srv.Entity.Addresses[0] {
-				conf, err := ca.NetworkSign(s)
-				if err != nil {
-					return err
-				}
-				dbg.Lvl3("New config:", conf)
-				ca.Config = conf
-			}
-		}
-	}
-	dbg.Lvl3(ca.Config.Servers)
-	return nil
-}
-
 func serverCheck(c *cli.Context) {
 	err := ServerCheck(clientApp)
 	if err != nil {
 		dbg.Error(err)
 	} else {
-		dbg.Print("Correctly checkd servers")
+		dbg.Print("Correctly checked servers")
 	}
 }
 
-func ServerCheck(ca *ssh_ks.ClientApp) error {
-	if ca.Config == nil {
-		return errors.New("No config defined")
-	}
-	if len(ca.Config.Servers) == 0 {
-		return errors.New("No servers defined")
-	}
-	var cnf *ssh_ks.Config
-	for _, srv := range ca.Config.Servers {
-		cnf_, err := ca.NetworkGetConfig(srv)
-		dbg.ErrFatal(err)
-		if cnf != nil {
-			if bytes.Compare(cnf.Hash(), cnf_.Hash()) != 0 {
-				return errors.New("Hashes should be the same\n" +
-					"1st server has " + fmt.Sprintln(cnf_) +
-					"Server" + fmt.Sprint(srv) + "has" + fmt.Sprintln(cnf_))
-			} else {
-				cnf = cnf_
-			}
-		}
-	}
-	return nil
+func clientAdd(c *cli.Context) {
+	err := ClientAdd(clientApp, nil)
+	dbg.ErrFatal(err)
 }
 
-func clientAdd(c *cli.Context) {}
-func clientDel(c *cli.Context) {}
-func update(c *cli.Context)    {}
+func clientDel(c *cli.Context) {
+	err := ClientDel(clientApp, nil)
+	dbg.ErrFatal(err)
+}
+func update(c *cli.Context) {
+	dbg.ErrFatal(Update(clientApp, nil))
+	dbg.Print("Got latest configuration")
+}
